Make NGB Id a 3 letter country code
Note to the future self - when dealing with structs and FirstOrDefault - if you want to get null you need to make a Cast<T?>() first

Also Postgres doesn't differentiate varchar and nvarchar like SQL Server - great!

And when making migrations make sure to have them fail if current data is not up to standard.